requirements.txt
torch
streamlit
numpy
pillow
opencv-python
torchvision
ultralytics
pandas
matplotlib
tqdm
scikit-learn